requirements.txt
fastapi
uvicorn
PyPDF2
faiss-cpu
sentence-transformers
python-dotenv
python-multipart
requests
pytest
httpx
